# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Feed Ask AI API
  description: |
    A Retrieval Augmented Generation (RAG) system for PDF processing and intelligent Q&A.
    Upload PDF files, automatically embed them using Ollama models, and ask questions with AI-powered answers.
  version: 1.0.0
  contact:
    name: Feed Ask AI Support
  license:
    name: MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /api/v1/feed-ask/feed:
    post:
      tags:
        - Feed Management
      summary: Upload and Process PDF
      description: |
        Upload a PDF file for processing. The system will:
        1. Extract text from the PDF
        2. Split text into chunks (500 characters with 100 character overlap)
        3. Generate embeddings using nomic-embed-text model
        4. Store embeddings in MySQL database
        
        Previously uploaded vectors persist and can be queried together.
      operationId: uploadFeed
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to upload (max 100MB)
              required:
                - file
      responses:
        '200':
          description: PDF successfully uploaded and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedUploadResponse'
        '400':
          description: Invalid PDF or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File size exceeds maximum allowed (100MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during PDF processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Ollama service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/feed-ask/ask:
    get:
      tags:
        - Question Answering
      summary: Ask a Question
      description: |
        Ask a question based on previously uploaded PDF documents.
        The system will:
        1. Generate embedding for the question using nomic-embed-text
        2. Search for top 3 most relevant vector embeddings using cosine similarity
        3. Build context from matching document chunks
        4. Generate an intelligent answer using tinyllama model with the context
        
        Returns the answer along with source vectors and confidence scores.
      operationId: askQuestion
      parameters:
        - name: prompt
          in: query
          required: true
          description: The question to ask about the uploaded documents
          schema:
            type: string
            minLength: 1
            maxLength: 1000
          example: "What is the main topic of the document?"
      responses:
        '200':
          description: Question successfully answered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '400':
          description: Invalid question prompt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during question processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Ollama service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    FeedUploadResponse:
      type: object
      description: Response after successful PDF upload and processing
      properties:
        status:
          type: string
          enum:
            - success
          description: Status of the upload operation
          example: "success"
        message:
          type: string
          description: Descriptive message about the upload
          example: "PDF uploaded and processed successfully"
        data:
          type: object
          description: Upload details and metadata
          properties:
            documentId:
              type: integer
              format: int64
              description: Unique identifier for the uploaded PDF document
              example: 1
            documentName:
              type: string
              description: Name of the uploaded PDF file
              example: "document.pdf"
            pageCount:
              type: integer
              description: Number of pages in the PDF
              example: 5
            embeddingsCount:
              type: integer
              description: Number of text chunks that were embedded
              example: 12
            totalCharacters:
              type: integer
              description: Total characters extracted from the PDF
              example: 2500
            processingTimeMs:
              type: integer
              format: int64
              description: Time taken to process the PDF in milliseconds
              example: 8500
            uploadedAt:
              type: string
              format: date-time
              description: Timestamp when the file was uploaded
              example: "2024-05-04T10:30:00.000Z"
          required:
            - documentId
            - documentName
            - pageCount
            - embeddingsCount
            - totalCharacters
            - processingTimeMs
            - uploadedAt
      required:
        - status
        - message
        - data

    AskResponse:
      type: object
      description: Response with AI-generated answer and source attribution
      properties:
        status:
          type: string
          enum:
            - success
          description: Status of the question answering operation
          example: "success"
        prompt:
          type: string
          description: The original question prompt
          example: "What is this document about?"
        answer:
          type: string
          description: AI-generated answer based on document context
          example: "This document is about machine learning fundamentals including supervised and unsupervised learning."
        sourceVectors:
          type: array
          description: Top 3 most relevant document chunks used to generate the answer
          items:
            $ref: '#/components/schemas/VectorMatch'
          example:
            - sourceText: "Machine learning is a subset of artificial intelligence..."
              confidenceScore: 0.95
              pageNumber: 1
            - sourceText: "Supervised learning requires labeled training data..."
              confidenceScore: 0.87
              pageNumber: 2
            - sourceText: "Unsupervised learning discovers patterns in unlabeled data..."
              confidenceScore: 0.82
              pageNumber: 3
        metrics:
          type: object
          description: Performance metrics for the question answering process
          properties:
            questionEmbeddingTimeMs:
              type: integer
              format: int64
              description: Time to generate embedding for the question
              example: 1500
            vectorSearchTimeMs:
              type: integer
              format: int64
              description: Time to search for similar vectors
              example: 50
            answerGenerationTimeMs:
              type: integer
              format: int64
              description: Time to generate the answer using Ollama
              example: 5000
            totalProcessingTimeMs:
              type: integer
              format: int64
              description: Total time for the entire operation
              example: 6550
          required:
            - questionEmbeddingTimeMs
            - vectorSearchTimeMs
            - answerGenerationTimeMs
            - totalProcessingTimeMs
      required:
        - status
        - prompt
        - answer
        - sourceVectors
        - metrics

    VectorMatch:
      type: object
      description: A single matching vector document chunk with relevance score
      properties:
        sourceText:
          type: string
          description: The actual text from the document that matched
          example: "Machine learning is a subset of artificial intelligence that enables systems to learn and improve from experience."
        confidenceScore:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Cosine similarity score (0-1) indicating relevance
          example: 0.95
        pageNumber:
          type: integer
          description: Page number in the source PDF document
          example: 1
        documentId:
          type: integer
          format: int64
          description: ID of the source document
          example: 1
      required:
        - sourceText
        - confidenceScore
        - pageNumber
        - documentId

    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2024-05-04T10:30:00.000Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type or category
          example: "PDF Processing Error"
        message:
          type: string
          description: Detailed error message
          example: "Invalid PDF file format"
        path:
          type: string
          description: API endpoint path where the error occurred
          example: "/api/v1/feed-ask/feed"
      required:
        - timestamp
        - status
        - error
        - message
        - path

tags:
  - name: Feed Management
    description: Operations for uploading and managing PDF documents
  - name: Question Answering
    description: Operations for asking questions about uploaded documents